requirements.txt
absl-py==2.3.1
cffi==2.0.0
flatbuffers==25.12.19
mediapipe==0.10.31
numpy==2.2.6

opencv-python-headless==4.12.0.88
pillow==12.0.0
pycparser==2.23
sounddevice==0.5.3

customtkinter==5.2.2
pygame==2.6.1
rembg==2.0.60
moviepy==1.0.3
edge-tts==7.2.7
librosa==0.11.0
